+++ js/data.js
@@ -1,0 +1,38 @@
+export const DESCRIPTIONS = [
+  'Закат на берегу океана',
+  'Котик отдыхает на подоконнике',
+  'Уличный музыкант в Париже',
+  'Горы на рассвете',
+  'Шумный вечер в центре города',
+  'Уютное кафе в Тбилиси',
+  'Дождливый день',
+  'Смешной момент с друзьями',
+  'Прогулка по лесу',
+  'Чашка кофе и хорошая книга',
+  'Радуга после грозы',
+  'Красивая архитектура старого города',
+  'Ужин на крыше',
+  'Пляж и коктейль',
+  'Утренняя йога на траве',
+  'Котёнок играет с клубком',
+  'Полет над облаками',
+  'Снежный вечер у камина',
+  'Рассвет в пустыне',
+  'Поездка в горы',
+  'Звёздное небо',
+  'Цветочный рынок',
+  'Круиз по Средиземному морю',
+  'Уличная еда в Азии',
+  'На вершине мира'
+];
+
+export const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+export const NAMES = ['Артём', 'Оля', 'Катя', 'Иван', 'Настя', 'Михаил', 'Дима', 'Алиса', 'Сергей', 'Лена'];
